+++ Changelog.md
@@ -6,3 +6,6 @@
 ## October 2022
 ### 15 October 2022
 * Updated Merlin and Eastin's Icons
+### 27 October 2022
+* Added Shield Hero character [Filo, Kizuna, Naofumi, Raphtalia]
+* Removed pointless help sidebar